requirements.txt
absl-py==2.3.1
accelerate==1.9.0
action-msgs==1.2.2
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
ament-cmake-test==1.3.12
ament-copyright==0.12.14
ament-cppcheck==0.12.14
ament-cpplint==0.12.14
ament-flake8==0.12.14
ament-index-python==1.4.0
ament-lint==0.12.14
ament-lint-cmake==0.12.14
ament-package==0.14.1
ament-pep257==0.12.14
ament-uncrustify==0.12.14
ament-xmllint==0.12.14
angles==1.15.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asciitree==0.3.3
async-timeout==5.0.1
attrs==25.3.0
av==15.0.0
beautifulsoup4==4.13.4
blinker==1.9.0
builtin-interfaces==1.2.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
cmake==4.0.3
composition-interfaces==1.2.2
cv-bridge==3.2.1
datasets==3.6.0
deepdiff==8.5.0
demo-nodes-py==0.20.5
diagnostic-msgs==4.9.0
diffusers==0.34.0
dill==0.3.8
dm-control==1.0.14
dm-env==1.6
dm-tree==0.1.9
docopt==0.6.2
domain-coordinator==0.10.0
draccus==0.10.0
einops==0.8.1
evdev==1.9.2
example-interfaces==0.9.3
examples-rclpy-executors==0.15.4
examples-rclpy-minimal-action-client==0.15.4
examples-rclpy-minimal-action-server==0.15.4
examples-rclpy-minimal-client==0.15.4
examples-rclpy-minimal-publisher==0.15.4
examples-rclpy-minimal-service==0.15.4
examples-rclpy-minimal-subscriber==0.15.4
Farama-Notifications==0.0.4
fasteners==0.19
feetech-servo-sdk==1.0.0
filelock==3.18.0
Flask==3.1.1
frozenlist==1.7.0
fsspec==2025.3.0
gdown==5.2.0
geometry-msgs==4.9.0
gitdb==4.0.12
GitPython==3.1.45
glfw==2.9.0
gym-aloha==0.1.1
gym-pusht==0.1.5
gymnasium==0.29.1
h5py==3.14.0
hf-xet==1.1.5
hf_transfer==0.1.9
huggingface-hub==0.34.3
idna==3.10
image-geometry==3.2.1
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
inquirerpy==0.3.4
interactive-markers==2.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonlines==4.0.0
labmaze==1.0.6
laser-geometry==2.4.0
launch==1.0.10
launch-ros==0.19.10
launch-testing==1.0.10
launch-testing-ros==0.19.10
launch-xml==1.0.10
launch-yaml==1.0.10
lazy_loader==0.4
-e git+https://github.com/huggingface/lerobot.git@71eff183ff1a286c9d3ad0962de3750a7ca00cb4#egg=lerobot
lifecycle-msgs==1.2.2
llvmlite==0.44.0
logging-demo==0.20.5
lxml==6.0.0
map-msgs==2.1.0
MarkupSafe==3.0.2
mergedeep==1.3.4
message-filters==4.3.8
mpmath==1.3.0
mujoco==2.3.7
multidict==6.6.3
multiprocess==0.70.16
mypy_extensions==1.1.0
nav-msgs==4.9.0
networkx==3.4.2
num2words==0.5.14
numba==0.61.2
numcodecs==0.13.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencv-python-headless==4.9.0.80
orderly-set==5.5.0
osrf-pycommon==2.1.6
packaging==25.0
pandas==2.3.1
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
pfzy==0.3.4
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pygame==2.6.1
pymunk==7.1.0
pynput==1.8.1
PyOpenGL==3.1.9
pyparsing==3.2.3
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-qt-binding==1.1.2
python-xlib==0.33
pytz==2025.2
PyYAML==6.0.2
pyyaml-include==1.4.1
pyzmq==27.0.0
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.5
rcl-interfaces==1.2.2
rclpy==3.3.17
rcutils==5.1.6
regex==2024.11.6
requests==2.32.4
rerun-sdk==0.22.1
resource-retriever==3.1.3
rmw-dds-common==1.6.0
ros2action==0.18.13
ros2bag==0.15.15
ros2cli==0.18.13
ros2component==0.18.13
ros2doctor==0.18.13
ros2interface==0.18.13
ros2launch==0.19.10
ros2lifecycle==0.18.13
ros2multicast==0.18.13
ros2node==0.18.13
ros2param==0.18.13
ros2pkg==0.18.13
ros2run==0.18.13
ros2service==0.18.13
ros2topic==0.18.13
rosbag2-interfaces==0.15.15
rosbag2-py==0.15.15
rosgraph-msgs==1.2.2
rosidl-adapter==3.1.7
rosidl-cli==3.1.7
rosidl-cmake==3.1.7
rosidl-generator-c==3.1.7
rosidl-generator-cpp==3.1.7
rosidl-generator-py==0.14.5
rosidl-parser==3.1.7
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.7
rosidl-typesupport-introspection-cpp==3.1.7
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.5
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.3
sensor-msgs==4.9.0
sensor-msgs-py==4.9.0
sentry-sdk==2.33.2
shape-msgs==4.9.0
shapely==2.1.1
six==1.17.0
smmap==5.0.2
soupsieve==2.7
sros2==0.10.7
statistics-msgs==1.2.2
std-msgs==4.9.0
std-srvs==4.9.0
stereo-msgs==4.9.0
sympy==1.13.1
teleop-twist-keyboard==2.4.0
termcolor==3.1.0
tf2-geometry-msgs==0.25.15
tf2-kdl==0.25.15
tf2-msgs==0.25.15
tf2-py==0.25.15
tf2-ros-py==0.25.15
tf2-sensor-msgs==0.25.15
tf2-tools==0.25.15
tifffile==2025.5.10
tokenizers==0.21.4
toml==0.10.2
topic-monitor==0.20.5
torch==2.6.0
TorchCodec==0.2.1
torchvision==0.21.0
tqdm==4.67.1
trajectory-msgs==4.9.0
transformers==4.51.3
triton==3.2.0
turtlesim==1.4.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
unique-identifier-msgs==2.2.1
urllib3==2.5.0
visualization-msgs==4.9.0
wandb==0.21.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zarr==2.18.3
zipp==3.23.0
